+++ sql/auth/sql_authorization.cc
@@ -2078,7 +2078,7 @@ err:
 
 bool check_grant_column(THD *thd, GRANT_INFO *grant,
                         const char *db_name, const char *table_name,
-                        const char *name, uint length,  Security_context *sctx)
+                        const char *name, size_t length,  Security_context *sctx)
 {
   GRANT_TABLE *grant_table;
   GRANT_COLUMN *grant_column;
@@ -2148,7 +2148,7 @@ err:
 */
 
 bool check_column_grant_in_table_ref(THD *thd, TABLE_LIST * table_ref,
-                                     const char *name, uint length)
+                                     const char *name, size_t length)
 {
   GRANT_INFO *grant;
   const char *db_name;
